Move fixture into conftest.py
