+++ tests/components/flow_it/test_fan.py
@@ -14,28 +14,10 @@ from homeassistant.components.fan import (
     SERVICE_TURN_OFF,
     SERVICE_TURN_ON,
 )
-from homeassistant.components.flow_it.const import DOMAIN
 from homeassistant.const import ATTR_ENTITY_ID, STATE_ON
 from homeassistant.core import HomeAssistant
 
 from tests.common import MockConfigEntry
-
-
-@pytest.fixture
-def mock_config_entry(hass: HomeAssistant) -> MockConfigEntry:
-    """Return a mock config entry."""
-    entry = MockConfigEntry(
-        domain=DOMAIN,
-        title="Flow-it Device",
-        unique_id="00:11:22:33:44:55",
-        data={
-            "host": "http://1.1.1.1",
-            "username": "api",
-            "password": "test-password",
-        },
-    )
-    entry.add_to_hass(hass)
-    return entry
 
 
 @pytest.mark.parametrize(
